requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
jinja2>=3.1.0

# HTTP scanner dependencies (required for full functionality)
httpx>=0.25.0
dnspython>=2.4.0
